+++ test.py
@@ -20,10 +20,15 @@ with pwncat.manager.Manager("data/pwncatrc") as manager:
     # session = manager.create_session("windows", host="192.168.122.11", port=4444)
     # session = manager.create_session("linux", host="pwncat-ubuntu", port=4444)
     # session = manager.create_session("linux", host="127.0.0.1", port=4444)
-    session = manager.create_session(
-        "linux", certfile="/tmp/cert.pem", keyfile="/tmp/cert.pem", port=4444
-    )
+    # session = manager.create_session(
+    #     "linux", certfile="/tmp/cert.pem", keyfile="/tmp/cert.pem", port=4444
+    # )
 
     # session.platform.powershell("amsiutils")
+
+    listener = manager.create_listener(
+        protocol="socket", host="0.0.0.0", port=4444, platform="windows"
+    )
+    listener = manager.create_listener(protocol="socket", host="0.0.0.0", port=9999)
 
     manager.interactive()