+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and simply didn't have the time to go back and retroactively create one.
 - Added Enumeration States to allow session-bound enumerations
 - Added PyPi publishing to GitHub `publish` workflow.
 - Added licensing for pwncat (MIT)
+- Added background listener API and commands ([#43](https://github.com/calebstewart/pwncat/issues/43))
+- Added Windows privilege escalation via BadPotato plugin ([#106](https://github.com/calebstewart/pwncat/issues/106))
 ### Removed
 - Removed `setup.py` and `requirements.txt`